add query to url

+++ staff/groups/ghibli-group/ghibli-app/components/app.jsx
@@ -6,45 +6,72 @@ class App extends Component {
     state= {view: undefined, error: undefined,query: undefined, token: undefined, results: undefined, category: undefined, result: undefined, loggedIn: false, toggleMenu: false, user: undefined, favs: undefined, message: undefined, resultsFilms: undefined, resultsPeople: undefined, resultsLocations: undefined, resultsSpecies: undefined, resultsVehicles: undefined,linkedFilms: undefined, linkedCharacters: undefined, linkedLocations: undefined, linkedSpecies: undefined, linkedVehicles: undefined}
         
     componentWillMount() {
-        const { token } = sessionStorage
+        
+        const { token } = sessionStorage       
 
-        if (token)
-            try {
-                retrieveUser(token, (error, user) => {
-                    if (error) {
-                        this.__handleError__(error)
-                        this.handleLogout()
-                    }
+        if(address.search.query){
+            const {query} = address.search
 
-                    if (user !== undefined) this.setState({ view: 'home', user, loggedIn: true })
-
-                    if(address.search.q){
-                        
-                        const {q: query} = address.search
+            if (token && query){
+                try {
+                    retrieveUser(token, (error, user) => {
+                        if (error) {
+                            this.__handleError__(error)
+                            this.handleLogout()
+                        }
     
+                        if (user !== undefined) this.setState({ user, loggedIn: true })
                         this.handleResults(query)
+                    })
                     
-                    }else if(address.hash && address.hash.startsWith(`${this.state.category}/`)){
-                        const [,id] = address.hash.split('/')
+                
+                } catch (error) {
+                    this.__handleError__(error)
+                  
+                    sessionStorage.clear()
     
-                        this.handleDetail(id, this.state.category)
-                    }
-                })
-            
-            } catch (error) {
-                this.__handleError__(error)
-              
-                sessionStorage.clear()
-
-                this.setState({ view: 'start' })
+                    this.setState({ view: 'start' })
+                }
+            }else if(query){
+                this.handleResults(query)
             }
-        else {
-            this.setState({view: 'start'})
+            
+        
+        }else if(address.hash && address.hash.startsWith(`${this.state.category}/`)){
+            const [,id] = address.hash.split('/')
 
-            setTimeout(() => {
-                this.handleGoToHome()
-            }, 1500)
-    }
+            this.handleDetail(id, this.state.category)
+        }else{
+            if (token){
+                try {
+                    retrieveUser(token, (error, user) => {
+                        if (error) {
+                            this.__handleError__(error)
+                            this.handleLogout()
+                        }
+    
+                        if (user !== undefined) this.setState({ view: 'home', user, loggedIn: true })
+                    })
+                    
+                
+                } catch (error) {
+                    this.__handleError__(error)
+                  
+                    sessionStorage.clear()
+    
+                    this.setState({ view: 'start' })
+                }
+            }
+            else {
+                this.setState({view: 'start'})
+
+                setTimeout(() => {
+                    this.handleGoToHome()
+                }, 1500)
+            }
+
+        }
+        
 }
  
     __handleError__(error) {
@@ -155,6 +182,10 @@ class App extends Component {
     handleResults = (query) => {
 
         const _query = toProperCase(query)
+
+        this.setState({query})
+
+        address.search = {query}
 
         searchFilms(_query, undefined, undefined, (error, resultsFilms) => {
             if(error)
@@ -296,7 +327,7 @@ class App extends Component {
 
     render() {
 
-        const {props: {title, query}, state: {view, error, results, category, result, user, resultsFilms, resultsPeople, resultsLocations, resultsSpecies, resultsVehicles, loggedIn, toggleMenu, message, linkedFilms, favs, linkedCharacters, linkedLocations, linkedSpecies, linkedVehicles}, handleGoToHome, handleGoToLogin, 
+        const {props: {title}, state: {view, query, error, results, category, result, user, resultsFilms, resultsPeople, resultsLocations, resultsSpecies, resultsVehicles, loggedIn, toggleMenu, message, linkedFilms, favs, linkedCharacters, linkedLocations, linkedSpecies, linkedVehicles}, handleGoToHome, handleGoToLogin, 
         handleResults, handleToggleMenu, handleGoToWatchlist, handleGoToEditProfile, handleLogout, handleUpdate, handleDeleteUser,
         handleLogin, handleRegister, handleGoToRegister, handleSearchFilms, handleSearch, handleSearchCategories, handleFav,
         handleDetail} = this
@@ -304,7 +335,7 @@ class App extends Component {
         return <main className="main">
             {view === "start" && <Init title={title} goToLanding={handleGoToHome}/>}
 
-            {view !== "start" && <Header goToLogin={handleGoToLogin} search={handleResults} goHome={handleGoToHome} showNav={handleToggleMenu} toggleMenu={toggleMenu} loggedIn={loggedIn} 
+            {view !== "start" && <Header query={query} goToLogin={handleGoToLogin} search={handleResults} goHome={handleGoToHome} showNav={handleToggleMenu} toggleMenu={toggleMenu} loggedIn={loggedIn} 
             //onSubmit={handleSearchFilms} 
             warning={error} goToWatchlist={handleGoToWatchlist} goToEditProfile={handleGoToEditProfile} logout={handleLogout} user={user}/>}
